Make fans move only when we get it right
It's way too much fun to just sit and stare at the fans moving, so let's make them only move when we get the answer right.

To do that I'm introducing the idea of different "celebration" configs. In this case we want to swap the image (to an animated gif). In other cases we'll stay with the default celebration.